refactor(arena): move schemas to dedicated module; squash branch migrations
- Arena request/response models moved from the endpoint into
  app/schemas/arena.py (router now imports them, matching the repo's
  schema-per-module convention)
- Squash the seven linear arena-branch migrations into one
  (9f8e7d6c5b4a3) anchored at master's head, so the branch's DB story is
  a single reviewable step instead of a long migration chain
