--- backend/app/alembic/versions/2026_08_23_0002-b3c2d1e0f9a8_add_incident_combat_progress.py
@@ -1,15 +1,0 @@
-"""Add combat_progress to incident for fractional kill accumulation."""
-
-from alembic import op
-
-revision = "b3c2d1e0f9a8"
-down_revision: str | None = "a2b1c9d8e7f6"
-branch_labels = depends_on = None
-
-
-def upgrade() -> None:
-    op.execute("ALTER TABLE incident ADD COLUMN combat_progress DOUBLE PRECISION NOT NULL DEFAULT 0")
-
-
-def downgrade() -> None:
-    op.execute("ALTER TABLE incident DROP COLUMN combat_progress")
